feat(node): Split up http integration into composable parts (#17524)
This is a first step to de-compose the node/node-core `httpIntegration`
into multiple composable parts:

* `httpServerIntegration` - handles request isolation, sessions, trace
continuation, so core Sentry functionality
* `httpServerSpansIntegration` - emits `http.server` spans for incoming
requests

The `httpIntegration` sets these up under the hood, and for now it is
not really recommended for users to use these stand-alone (though it is
possible if users opt-out of the `httpIntegration`). The reason is to
remain backwards compatible with users using/customizing the
`httpIntegration`. We can revisit this in a later major.

These new integrations have a much slimmer API surface, and also allows
us to avoid having to prefix all the options etc. with what they are
about (e.g. `incomingXXX` or `outgoingXXX`). It also means you can
actually tree-shake certain features (span creation) out, in theory.

Outgoing request handling remains the same for the time being, once we
decoupled this from the otel http instrumentation we can do something
similar there.

The biggest challenge was how to make it possible to de-compose this
without having to monkey patch the http server twice. I opted to allow
to add callbacks to the `httpServerIntegration` which it will call on
any request. So the `httpServerSpansIntegration` can register a callback
for itself and plug into this with little overhead.
