+++ packages/node-core/src/integrations/http/SentryHttpInstrumentation.ts
@@ -2,16 +2,15 @@ import type { ChannelListener } from 'node:diagnostics_channel';
 import { subscribe, unsubscribe } from 'node:diagnostics_channel';
 import type * as http from 'node:http';
 import type * as https from 'node:https';
-import type { Span } from '@opentelemetry/api';
 import { context } from '@opentelemetry/api';
 import { isTracingSuppressed } from '@opentelemetry/core';
 import type { InstrumentationConfig } from '@opentelemetry/instrumentation';
 import { InstrumentationBase, InstrumentationNodeModuleDefinition } from '@opentelemetry/instrumentation';
+import type { Span } from '@sentry/core';
 import { debug, LRUMap, SDK_VERSION } from '@sentry/core';
 import { DEBUG_BUILD } from '../../debug-build';
 import { getRequestUrl } from '../../utils/getRequestUrl';
 import { INSTRUMENTATION_NAME } from './constants';
-import { instrumentServer } from './incoming-requests';
 import {
   addRequestBreadcrumb,
   addTracePropagationHeadersToOutgoingRequest,
@@ -23,30 +22,11 @@ type Https = typeof https;
 
 export type SentryHttpInstrumentationOptions = InstrumentationConfig & {
   /**
-   * Whether breadcrumbs should be recorded for requests.
+   * Whether breadcrumbs should be recorded for outgoing requests.
    *
    * @default `true`
    */
   breadcrumbs?: boolean;
-
-  /**
-   * Whether to create spans for requests or not.
-   * As of now, creates spans for incoming requests, but not outgoing requests.
-   *
-   * @default `true`
-   */
-  spans?: boolean;
-
-  /**
-   * Whether to extract the trace ID from the `sentry-trace` header for incoming requests.
-   * By default this is done by the HttpInstrumentation, but if that is not added (e.g. because tracing is disabled, ...)
-   * then this instrumentation can take over.
-   *
-   * @deprecated This is always true and the option will be removed in the future.
-   *
-   * @default `true`
-   */
-  extractIncomingTraceFromHeader?: boolean;
 
   /**
    * Whether to propagate Sentry trace headers in outgoing requests.
@@ -56,20 +36,6 @@ export type SentryHttpInstrumentationOptions = InstrumentationConfig & {
    * @default `false`
    */
   propagateTraceInOutgoingRequests?: boolean;
-
-  /**
-   * Whether to automatically ignore common static asset requests like favicon.ico, robots.txt, etc.
-   * This helps reduce noise in your transactions.
-   *
-   * @default `true`
-   */
-  ignoreStaticAssets?: boolean;
-
-  /**
-   * If true, do not generate spans for incoming requests at all.
-   * This is used by Remix to avoid generating spans for incoming requests, as it generates its own spans.
-   */
-  disableIncomingRequestSpans?: boolean;
 
   /**
    * Do not capture breadcrumbs for outgoing HTTP requests to URLs where the given callback returns `true`.
@@ -82,55 +48,51 @@ export type SentryHttpInstrumentationOptions = InstrumentationConfig & {
    */
   ignoreOutgoingRequests?: (url: string, request: http.RequestOptions) => boolean;
 
+  // All options below do not do anything anymore in this instrumentation, and will be removed in the future.
+  // They are only kept here for backwards compatibility - the respective functionality is now handled by the httpServerIntegration/httpServerSpansIntegration.
+
   /**
-   * Do not capture spans for incoming HTTP requests to URLs where the given callback returns `true`.
-   *
-   * @param urlPath Contains the URL path and query string (if any) of the incoming request.
-   * @param request Contains the {@type IncomingMessage} object of the incoming request.
+   * @deprecated This no longer does anything.
+   */
+  spans?: boolean;
+
+  /**
+   * @depreacted This no longer does anything.
+   */
+  extractIncomingTraceFromHeader?: boolean;
+
+  /**
+   * @deprecated This no longer does anything.
+   */
+  ignoreStaticAssets?: boolean;
+
+  /**
+   * @deprecated This no longer does anything.
+   */
+  disableIncomingRequestSpans?: boolean;
+
+  /**
+   * @deprecated This no longer does anything.
    */
   ignoreSpansForIncomingRequests?: (urlPath: string, request: http.IncomingMessage) => boolean;
 
   /**
-   * Do not capture the request body for incoming HTTP requests to URLs where the given callback returns `true`.
-   * This can be useful for long running requests where the body is not needed and we want to avoid capturing it.
-   *
-   * @param url Contains the entire URL, including query string (if any), protocol, host, etc. of the incoming request.
-   * @param request Contains the {@type RequestOptions} object used to make the incoming request.
+   * @deprecated This no longer does anything.
    */
   ignoreIncomingRequestBody?: (url: string, request: http.RequestOptions) => boolean;
 
   /**
-   * A hook that can be used to mutate the span for incoming requests.
-   * This is triggered after the span is created, but before it is recorded.
-   */
-  incomingRequestSpanHook?: (span: Span, request: http.IncomingMessage, response: http.ServerResponse) => void;
-
-  /**
-   * Controls the maximum size of incoming HTTP request bodies attached to events.
-   *
-   * Available options:
-   * - 'none': No request bodies will be attached
-   * - 'small': Request bodies up to 1,000 bytes will be attached
-   * - 'medium': Request bodies up to 10,000 bytes will be attached (default)
-   * - 'always': Request bodies will always be attached
-   *
-   * Note that even with 'always' setting, bodies exceeding 1MB will never be attached
-   * for performance and security reasons.
-   *
-   * @default 'medium'
+   * @deprecated This no longer does anything.
    */
   maxIncomingRequestBodySize?: 'none' | 'small' | 'medium' | 'always';
 
   /**
-   * Whether the integration should create [Sessions](https://docs.sentry.io/product/releases/health/#sessions) for incoming requests to track the health and crash-free rate of your releases in Sentry.
-   * Read more about Release Health: https://docs.sentry.io/product/releases/health/
-   *
-   * Defaults to `true`.
+   * @deprecated This no longer does anything.
    */
   trackIncomingRequestsAsSessions?: boolean;
 
   /**
-   * @deprecated This is deprecated in favor of `incomingRequestSpanHook`.
+   * @deprecated This no longer does anything.
    */
   instrumentation?: {
     requestHook?: (span: Span, req: http.ClientRequest | http.IncomingMessage) => void;
@@ -143,9 +105,7 @@ export type SentryHttpInstrumentationOptions = InstrumentationConfig & {
   };
 
   /**
-   * Number of milliseconds until sessions tracked with `trackIncomingRequestsAsSessions` will be flushed as a session aggregate.
-   *
-   * Defaults to `60000` (60s).
+   * @deprecated This no longer does anything.
    */
   sessionFlushingDelayMS?: number;
 };
@@ -180,24 +140,6 @@ export class SentryHttpInstrumentation extends InstrumentationBase<SentryHttpIns
     // but we only want to register them once, whichever is loaded first
     let hasRegisteredHandlers = false;
 
-    const spansEnabled = this.getConfig().spans ?? true;
-
-    const onHttpServerRequestStart = ((_data: unknown) => {
-      const data = _data as { server: http.Server };
-      instrumentServer(data.server, {
-        // eslint-disable-next-line deprecation/deprecation
-        instrumentation: this.getConfig().instrumentation,
-        ignoreIncomingRequestBody: this.getConfig().ignoreIncomingRequestBody,
-        ignoreSpansForIncomingRequests: this.getConfig().ignoreSpansForIncomingRequests,
-        incomingRequestSpanHook: this.getConfig().incomingRequestSpanHook,
-        maxIncomingRequestBodySize: this.getConfig().maxIncomingRequestBodySize,
-        trackIncomingRequestsAsSessions: this.getConfig().trackIncomingRequestsAsSessions,
-        sessionFlushingDelayMS: this.getConfig().sessionFlushingDelayMS ?? 60_000,
-        ignoreStaticAssets: this.getConfig().ignoreStaticAssets,
-        spans: spansEnabled && !this.getConfig().disableIncomingRequestSpans,
-      });
-    }) satisfies ChannelListener;
-
     const onHttpClientResponseFinish = ((_data: unknown) => {
       const data = _data as { request: http.ClientRequest; response: http.IncomingMessage };
       this._onOutgoingRequestFinish(data.request, data.response);
@@ -220,7 +162,6 @@ export class SentryHttpInstrumentation extends InstrumentationBase<SentryHttpIns
 
       hasRegisteredHandlers = true;
 
-      subscribe('http.server.request.start', onHttpServerRequestStart);
       subscribe('http.client.response.finish', onHttpClientResponseFinish);
 
       // When an error happens, we still want to have a breadcrumb
@@ -238,7 +179,6 @@ export class SentryHttpInstrumentation extends InstrumentationBase<SentryHttpIns
     };
 
     const unwrap = (): void => {
-      unsubscribe('http.server.request.start', onHttpServerRequestStart);
       unsubscribe('http.client.response.finish', onHttpClientResponseFinish);
       unsubscribe('http.client.request.error', onHttpClientRequestError);
       unsubscribe('http.client.request.created', onHttpClientRequestCreated);
