+++ packages/node/src/index.ts
@@ -157,6 +157,8 @@ export type {
 
 export {
   logger,
+  httpServerIntegration,
+  httpServerSpansIntegration,
   nodeContextIntegration,
   contextLinesIntegration,
   localVariablesIntegration,
