+++ packages/remix/src/server/index.ts
@@ -48,6 +48,8 @@ export {
   graphqlIntegration,
   hapiIntegration,
   httpIntegration,
+  httpServerIntegration,
+  httpServerSpansIntegration,
   // eslint-disable-next-line deprecation/deprecation
   inboundFiltersIntegration,
   eventFiltersIntegration,
